requirements.txt
Click==7.0
Flask==1.0.2
Flask-Markdown==0.3
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.1
Pillow==5.4.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-slugify==2.0.1
six==1.12.0
SQLAlchemy==1.2.18
Unidecode==1.0.23
Werkzeug==0.14.1
WTForms==2.2.1